+++ GenshinWish/Models/DTO/WishDetailDto.cs
@@ -10,15 +10,15 @@ namespace GenshinWish.Models.DTO
         public int Star3Count { get; set; }
         public int Star4Count { get; set; }
         public int Star5Count { get; set; }
-        public int CharStar4Rate { get; set; }
-        public int WpnStar4Rate { get; set; }
-        public int StdStar4Rate { get; set; }
-        public int CharStar5Rate { get; set; }
-        public int WpnStar5Rate { get; set; }
-        public int StdStar5Rate { get; set; }
-        public int CharWishTimes { get; set; }
-        public int WpnWishTimes { get; set; }
-        public int StdWishTimes { get; set; }
+        public decimal CharStar4Rate { get; set; }
+        public decimal WpnStar4Rate { get; set; }
+        public decimal StdStar4Rate { get; set; }
+        public decimal CharStar5Rate { get; set; }
+        public decimal WpnStar5Rate { get; set; }
+        public decimal StdStar5Rate { get; set; }
+        public decimal CharWishTimes { get; set; }
+        public decimal WpnWishTimes { get; set; }
+        public decimal StdWishTimes { get; set; }
         public int TotalWishTimes { get; set; }
     }
 }
